# contracts/openapi/core-api.v1.yaml
openapi: 3.0.3
info:
  title: HospitalAI Core API
  version: 1.0.0
paths:
...
  /api/workbench/{encounterId}:
    get:
      summary: Build patient context, deterministic safety rules, controlled evidence, and candidates.
      parameters:
        - name: encounterId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Workbench payload
  /api/recommendations/{recommendationId}/decision:
    post:
      summary: Adopt, modify, or reject a recommendation and create a simulated HIS draft when allowed.
      parameters:
        - name: recommendationId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Persisted decision and draft status
...
  /api/integration/his/snapshots/import:
    post:
      summary: Import a versioned HIS snapshot into the unified medical model.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceSystem, sourceBatchId, payload]
              properties:
                sourceSystem: { type: string }
                sourceBatchId: { type: string }
                payload: { type: object }
      responses:
        "200":
          description: Import event status and upsert counts.